@@ -1,12 +1,11 @@
 package de.geheimagentnr1.auto_restart;
 
 import de.geheimagentnr1.auto_restart.config.ServerConfig;
-import net.minecraftforge.fml.ExtensionPoint;
+import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.fml.network.FMLNetworkConstants;
-import org.apache.commons.lang3.tuple.Pair;
+import net.minecraftforge.fmllegacy.network.FMLNetworkConstants;
 
 
 @SuppressWarnings( "UtilityClassWithPublicConstructor" )
@@ -20,8 +19,11 @@ public class AutoRestart {
 		
 		ModLoadingContext.get().registerConfig( ModConfig.Type.SERVER, ServerConfig.CONFIG );
 		ModLoadingContext.get().registerExtensionPoint(
-			ExtensionPoint.DISPLAYTEST,
-			() -> Pair.of( () -> FMLNetworkConstants.IGNORESERVERONLY, ( remote, isServer ) -> true )
+			IExtensionPoint.DisplayTest.class,
+			() -> new IExtensionPoint.DisplayTest(
+				() -> FMLNetworkConstants.IGNORESERVERONLY,
+				( remote, isServer ) -> true
+			)
 		);
 	}
 }
